many updates including i18next support init
