+++ src/components/CommitteeTools/components/KeysSidebar/index.scss
@@ -1,0 +1,1 @@
+@import "../../../../styles/constants.scss";
