+++ src/components/CommitteeTools/index.scss
@@ -1,0 +1,12 @@
+@import "../../styles/constants.scss";
+
+.committee-tools-wrapper {
+  display: flex;
+  flex-direction: row;
+}
+
+.keys-navbar {
+  min-width: 180px;
+  display: flex;
+  flex-direction: column;
+}
