--- src/components/CommitteeTools/style.scss
@@ -1,13 +1,0 @@
-@import "../../styles/constants.scss";
-
-.committee-tools-wrapper {
-    display: flex;
-    flex-direction: row;
-}
-
-
-.keys-navbar {
-    min-width: 180px;
-    display: flex;
-    flex-direction: column;
-}
